refactor tool_calling_llm.py
