+++ holmes/core/tool_calling_llm.py
@@ -3,32 +3,28 @@ import json
 import logging
 import textwrap
 import os
-from typing import Dict, Generator, List, Optional
+from typing import Dict, Generator, List, Optional, Iterator
 
 import litellm
 import jinja2
 from openai import BadRequestError, OpenAI
 from openai._types import NOT_GIVEN
 from openai.types.chat.chat_completion import Choice
+from openai.types.chat.chat_completion_message_tool_call import ChatCompletionMessageToolCall
 from openai.types.chat.chat_completion_chunk import ChatCompletionChunk
 from pydantic import BaseModel
 from rich.console import Console
 
 from holmes.core.issue import Issue
 from holmes.core.runbooks import RunbookManager
-from holmes.core.tools import YAMLToolExecutor
+from holmes.core.tools import YAMLToolExecutor, ToolCallResult
 
-
-class ToolCallResult(BaseModel):
-    tool_name: str
-    description: str
-    result: str
 
 class LLMResult(BaseModel):
     tool_calls: Optional[List[ToolCallResult]] = None
     result: Optional[str] = None
 
-    # TODO: clean up these two
+    # TODO: clean up these two - prompt is really messages and messages is never set
     prompt: Optional[str] = None
     messages: Optional[List[dict]] = None
 
@@ -71,9 +67,52 @@ class ToolCallingLLM:
         # this unfortunately does not seem to work for azure if the deployment name is not a well-known model name 
         #if not litellm.supports_function_calling(model=model):
         #    raise Exception(f"model {model} does not support function calling. You must use HolmesGPT with a model that supports function calling.")
-
+    
     def call(self, system_prompt, user_prompt) -> LLMResult:
-        messages = [
+        messages = self._get_initial_messages(system_prompt, user_prompt)
+        tool_calls = []
+        tools = self.tool_executor.get_all_tools_openai_format()
+        for i in range(self.max_steps):
+            logging.debug(f"running iteration {i}")
+            # on the last step we don't allow tools - we want to force a reply, not a request to run another tool
+            tools = NOT_GIVEN if i == self.max_steps - 1 else tools
+            tool_choice = NOT_GIVEN if tools == NOT_GIVEN else "auto"
+            logging.debug(f"sending messages {messages}")
+            full_response = litellm.completion(
+                model=self.model,
+                api_key=self.api_key,
+                messages=messages,
+                tools=tools,
+                tool_choice=tool_choice,
+                temperature=0.00000001,
+            )
+            response = full_response.choices[0]
+            response_message = response.message
+            messages.append(
+                response_message.model_dump(
+                    exclude_defaults=True, exclude_unset=True, exclude_none=True
+                )
+            )
+
+            tools_to_call: Optional[List[ChatCompletionMessageToolCall]]  = getattr(response_message, "tool_calls", None)
+            if not tools_to_call:
+                return LLMResult(
+                    result=response_message.content,
+                    tool_calls=tool_calls,
+                    prompt=json.dumps(messages, indent=2),
+                )
+
+            # Some models return text responses alongside tool calls
+            if response_message.content:
+                logging.info(f"AI: {response_message.content}")
+
+            tool_results = self.tool_executor.invoke_tools_openai_format(tools_to_call)
+            tool_calls.extend(tool_results)
+            messages.extend([tool_result.to_openai_format() for tool_result in tool_results])
+
+        
+    def _get_initial_messages(self, system_prompt, user_prompt):
+        return [
             {
                 "role": "system",
                 "content": system_prompt,
@@ -83,76 +122,7 @@ class ToolCallingLLM:
                 "content": user_prompt,
             },
         ]
-        tool_calls = []
-        tools = self.tool_executor.get_all_tools_openai_format()
-        for i in range(self.max_steps):
-            logging.debug(f"running iteration {i}")
-            # on the last step we don't allow tools - we want to force a reply, not a request to run another tool
-            tools = NOT_GIVEN if i == self.max_steps - 1 else tools
-            tool_choice = NOT_GIVEN if tools == NOT_GIVEN else "auto"
-            logging.debug(f"sending messages {messages}")
-            try:
-                full_response = litellm.completion(
-                    model=self.model,
-                    api_key=self.api_key,
-                    messages=messages,
-                    tools=tools,
-                    tool_choice=tool_choice,
-                    temperature=0.00000001
-                )
-                logging.debug(f"got response {full_response}")
-            # catch a known error that occurs with Azure and replace the error message with something more obvious to the user
-            except BadRequestError as e:
-                if "Unrecognized request arguments supplied: tool_choice, tools" in str(e):
-                    raise Exception(
-                        "The Azure model you chose is not supported. Model version 1106 and higher required."
-                    )
-                else:
-                    raise
-            response = full_response.choices[0]
-            response_message = response.message
-            messages.append(
-                response_message.model_dump(
-                    exclude_defaults=True, exclude_unset=True, exclude_none=True
-                )
-            )
-
-            tools_to_call = getattr(response_message, "tool_calls", None)
-            if not tools_to_call:
-                return LLMResult(
-                    result=response_message.content,
-                    tool_calls=tool_calls,
-                    prompt=json.dumps(messages, indent=2),
-                )
-
-            # when asked to run tools, we expect no response other than the request to run tools
-            if response_message.content:
-                logging.warning(f"got unexpected response when tools were given: {response_message.content}")
-
-            for t in tools_to_call:
-                tool_name = t.function.name
-                tool_params = json.loads(t.function.arguments)
-                tool = self.tool_executor.get_tool_by_name(tool_name)
-                tool_response = tool.invoke(tool_params)
-                MAX_CHARS = 100_000 # an arbitrary limit - we will do something smarter in the future
-                if len(tool_response) > MAX_CHARS:
-                    logging.warning(f"tool {tool_name} returned a very long response ({len(tool_response)} chars) - truncating to last 10000 chars")
-                    tool_response = tool_response[-MAX_CHARS:]
-                messages.append(
-                    {
-                        "tool_call_id": t.id,
-                        "role": "tool",
-                        "name": tool_name,
-                        "content": tool_response,
-                    }
-                )
-                tool_calls.append(
-                    ToolCallResult(
-                        tool_name=tool_name,
-                        description=tool.get_parameterized_one_liner(tool_params),
-                        result=tool_response,
-                    )
-                )
+    
 
 # TODO: consider getting rid of this entirely and moving templating into the cmds in holmes.py 
 class IssueInvestigator(ToolCallingLLM):
